Add collection products feature with navigation
Implement collection products viewing with navigation from categories and brands lists. Added CollectionProductsView and CollectionProductsViewModel to display products from a selected collection. Extended data layer with fetchCollectionProducts endpoint. Refactored ProductCard and ProductsGrid components to accept data and callbacks directly, reducing tight coupling to HomeViewModel. Fixed Firebase initialization to check if app is already configured.

+++ Buyza App/Buyza App/Presentation/Home/ViewModel/HomeViewModel.swift
@@ -61,9 +61,9 @@ class HomeViewModel : ObservableObject {
         }
     }
     
-    func isFavorite(productID: Int) -> Bool {
+    func isFavorite(productID: Int64) -> Bool {
         // will call here the function from the usecase
-        
+
         return true
     }
 }
